Re-introduced ledger protocol. Now as short of job portal for clerks.
As it was not good to that every clerk would have their own mailbox I created a centralised version of it.

Other than that moved test.txt out from sysbin because it was causing a directory update problem? Not enough room in the dir? I suspect it is because shell is located there.

 then I cleaned a bit and wrote comments.

+++ kernel/fs/fat32/fat32main.c
@@ -18,7 +18,7 @@ fat32_fs_t f32_fs;
  * Return: STATUS_ERROR || STATUS_OK.
  */
 int fat32_find_cluster(const char *path, uint32_t *out_cluster, uint32_t *out_size,
-    char *out_fname, uint32_t *out_attr) {
+    char *out_fname, uint8_t *out_attr) {
 
     if (path == NULL) {
         ERROR("[FAT32][FIND_CLUSTER]: Given path was invalid\n");
